Rename ext.kt in the package char to char.ext.kt Move mapper.ext.kt file to util folder.
